Implemented transaction swipe to delete

+++ app/src/main/java/com/mancode/financetracker/viewmodel/TransactionViewModel.kt
@@ -5,7 +5,8 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import com.mancode.financetracker.database.entity.TransactionEntity
 import com.mancode.financetracker.database.pojos.TransactionFull
-import com.mancode.financetracker.repository.DataRepository
+import com.mancode.financetracker.repository.TransactionsRepository
+import com.mancode.financetracker.utils.InjectorUtils
 
 /**
  * Created by Manveru on 12.02.2018.
@@ -13,11 +14,19 @@ import com.mancode.financetracker.repository.DataRepository
 
 class TransactionViewModel(application: Application) : AndroidViewModel(application) {
 
-    private val repository: DataRepository = DataRepository.getInstance(application)
+    private val repository: TransactionsRepository = InjectorUtils.getTransactionsRepository(application)
     val allTransactions: LiveData<List<TransactionFull>> by lazy { repository.allTransactions }
 
-    fun deleteTransaction(transaction: TransactionEntity) {
-        repository.deleteTransaction(transaction)
+    fun restoreTransaction(transaction: TransactionEntity?) {
+        if (transaction != null) {
+            repository.insertTransaction(transaction)
+        }
+    }
+
+    fun deleteTransaction(transaction: TransactionEntity?) {
+        if (transaction != null) {
+            repository.deleteTransaction(transaction)
+        }
     }
 
     fun toggleBookmark(transaction: TransactionEntity) {
